+++ examples/sync_consume.rs
@@ -1,4 +1,5 @@
 // paho-mqtt/examples/sync_consume.rs
+//
 // This is a Paho MQTT Rust client, sample application.
 //
 //! This application is an MQTT consumer/subscriber using the Rust
@@ -18,7 +19,7 @@
 //!
 
 /*******************************************************************************
- * Copyright (c) 2017-2022 Frank Pagliughi <fpagliughi@mindspring.com>
+ * Copyright (c) 2017-2023 Frank Pagliughi <fpagliughi@mindspring.com>
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -159,7 +160,6 @@ fn main() {
     // otherwise we're already disconnected.
     if cli.is_connected() {
         println!("\nDisconnecting...");
-        cli.unsubscribe_many(&subscriptions).unwrap();
         cli.disconnect(None).unwrap();
     }
     println!("Exiting");